Remove debugFBO
